+++ src/templates/location.tsx
@@ -10,7 +10,7 @@ import {
 } from "@yext/pages";
 import { DocumentProvider } from "@yext/pages/util";
 import { Config, Render } from "@measured/puck";
-import { locationConfig } from "../puck/puck.config";
+import { locationConfig } from "../ve.config";
 import { resolveVisualEditorData } from "@yext/visual-editor";
 
 export const config: TemplateConfig = {
